@@ -22,12 +22,14 @@
   "devDependencies": {
     "@babel/core": "^7.15.0",
     "@types/d3": "^7.0.0",
+    "@vue/compiler-sfc": "^3.1.0",
     "babel-loader": "^8.2.2",
     "css-loader": "^6.2.0",
     "prettier": "2.3.2",
+    "style-loader": "^3.2.1",
     "ts-loader": "^9.2.5",
     "typescript": "^4.4.2",
-    "vue-loader": "^15.9.8",
+    "vue-loader": "^16.0.0",
     "webpack": "^5.51.1",
     "webpack-cli": "^4.8.0"
   },
@@ -35,7 +37,6 @@
     "axios": "^0.21.1",
     "d3": "^7.0.1",
     "ol": "^6.6.1",
-    "vue": "^2.6.14",
-    "vue-template-compiler": "^2.6.14"
+    "vue": "^3.1.0"
   }
 }
