@@ -13,7 +13,7 @@
 </template>
 
 <script lang="ts">
-import Vue from "vue";
+import { defineComponent } from "vue";
 import { JsonOccurrence } from "../interfaces";
 
 interface OccurrencesForDisplay {
@@ -26,7 +26,7 @@ interface OccurrencesForDisplay {
   url: string;
 }
 
-export default Vue.extend({
+export default defineComponent({
   name: "OccurrenceTablePage",
   computed: {
     preparedOccurrences: function (): OccurrencesForDisplay[] {
@@ -37,7 +37,7 @@ export default Vue.extend({
           lon: occ.lon ? occ.lon.toFixed(4) : "",
           date: occ.date,
           speciesName: occ.speciesName,
-          url: this.occurrencePageUrlTemplate.replace(
+          url: this.occurrencePageUrlTemplate!.replace(
             "{pk}",
             occ.id.toString()
           ),
