@@ -1,4 +1,6 @@
 declare module "*.vue" {
-  import Vue from "vue";
+  import { CompatVue } from "@vue/runtime-dom";
+  const Vue: CompatVue;
   export default Vue;
+  export * from "@vue/runtime-dom";
 }
